Corner Grocer App C++ Project 3
This commit adds the fundamental files for the Corner Grocer App, a C++ project developed for a comprehensive understanding of application development in a real-world context. The files are prioritized based on their contribution to the project's core functionality and structure:

`CornerGrocerApp.cpp`
The primary source file containing the complete implementation of the GroceryFrequency class and the main function. This file is crucial as it defines the application's functionality, handling data processing, user interactions, and output presentation.

`CS210_Project_Three_Input_File.txt`
The main input file with sample grocery item data. This file is essential for the application as it provides the real data the program processes, thereby demonstrating its practical application in analyzing purchasing patterns.

`frequency.dat`
A backup file that stores the frequency of each item processed by the application. Its role in ensuring data persistence is critical for maintaining data integrity and reliability.

`CornerGrocerApp.sln`
The solution file for the Visual Studio project, crucial for managing the overall project structure, dependencies, and configurations. It is the starting point for compiling and running the entire application.

`CornerGrocerApp.vcxproj` 
This project file is vital for defining project configurations, compiler settings, and linking the necessary libraries. It is a key component in the build process of the application.

`CornerGrocerApp.vcxproj.filters`
This file organizes the files in the project into filters (virtual folders), enhancing the project's navigability and manageability in Visual Studio. It plays a significant role in maintaining the project's structure and readability.

`CornerGrocerApp.vcxproj.user`
A user-specific project file that stores individual developer preferences and settings in Visual Studio. While not critical for the project's functionality, it personalizes the development experience.